requirements.txt
streamlit>=1.37.0,<2
streamlit-autorefresh>=1.0.1,<2
streamlit-authenticator>=0.4.1,<1
langchain>=0.2.16,<0.4
langchain-community>=0.2.16,<0.4
langchain-huggingface>=0.0.3,<1
langchain-groq>=0.1.9,<0.3
rank-bm25>=0.2.2,<1
sentence-transformers>=3.0.1,<4
python-dotenv>=1.0.1,<2
pypdf>=4.3.1,<5
bcrypt>=4.2.0,<5
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<1
python-jose[cryptography]>=3.3.0,<4
sqlalchemy>=2.0.36,<3
psycopg[binary]>=3.2.4,<4
pgvector>=0.3.6,<1
redis>=5.2.1,<6
minio>=7.2.13,<8
